requirements.txt
matplotlib==3.10.8
numpy==2.3.5
pandas==2.3.3
psycopg2_binary==2.9.11
Requests==2.32.5
